+++ backend/server/src/main/java/com/airbnb/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.airbnb.dao;
+package com.airbnb.repository;
 
 import com.airbnb.domain.User;
 import com.airbnb.exception.SqlException;
