+++ backend/server/src/main/resources/application.properties
@@ -5,6 +5,7 @@ github.client.secret=2def405b1649cdb94ac5c3aa430926f38cd1a242
 logging.level.sql=DEBUG
 logging.level.org.springframework.jdbc.core=DEBUG
 
+# MySQL DB 관련 설정
 spring.datasource.initialization-mode=always
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.schema=classpath:schema.sql
